refactor: 상수 사용하여 에러 메세지 처리
