+++ src/test/java/racingcar/validator/CarNameValidatorTest.java
@@ -1,7 +1,7 @@
-package racingcar.model;
+package racingcar.validator;
 
 import org.junit.jupiter.api.Test;
-import racingcar.validator.CarNameValidator;
+import racingcar.constant.ErrorMessage;
 
 import java.util.List;
 
@@ -12,27 +12,27 @@ class CarNameValidatorTest {
     void validateCarNames_should_throw_exception_when_name_length_exceeds_maximum() {
         assertThatThrownBy(() -> CarNameValidator.validateCarNames(List.of("pobi", "woni", "seobin")))
                 .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("자동차 이름은 1~5자 이하만 가능합니다.");
+                .hasMessage(ErrorMessage.CAR_NAME_LENGTH_ERROR);
     }
 
     @Test
     void validateCarNames_should_throw_exception_when_duplicate_names_exist() {
         assertThatThrownBy(() -> CarNameValidator.validateCarNames(List.of("pobi", "woni", "pobi")))
                 .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("자동차 이름은 중복될 수 없습니다.");
+                .hasMessage(ErrorMessage.CAR_NAME_DUPLICATE_ERROR);
     }
 
     @Test
     void validateCarNames_should_throw_exception_when_name_is_empty() {
         assertThatThrownBy(() -> CarNameValidator.validateCarNames(List.of("pobi", " ", "woni")))
                 .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("자동차 이름은 공백일 수 없습니다.");
+                .hasMessage(ErrorMessage.CAR_NAME_EMPTY_ERROR);
     }
 
     @Test
     void validateCarNames_should_throw_exception_when_name_is_not_alphabet() {
         assertThatThrownBy(() -> CarNameValidator.validateCarNames(List.of("pob1", "123")))
                 .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("자동차 이름은 알파벳만 가능합니다.");
+                .hasMessage(ErrorMessage.CAR_NAME_FORMAT_ERROR);
     }
 }
